perf: class for custom item events
