+++ src/main/java/net/eriagacha/register/RegisterItems.java
@@ -1,23 +1,15 @@
 package net.eriagacha.register;
 
 import lombok.extern.log4j.Log4j2;
-import net.eriagacha.network.NetworkClient;
-import net.eriagacha.utils.GachaUtils;
+import net.eriagacha.CustomItem;
 import net.eriagacha.utils.NameSpaces;
-import net.fabricmc.api.EnvType;
-import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.entity.effect.StatusEffectInstance;
 import net.minecraft.entity.effect.StatusEffects;
-import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.FoodComponent;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemGroup;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.Hand;
 import net.minecraft.util.Identifier;
-import net.minecraft.util.TypedActionResult;
 import net.minecraft.util.registry.Registry;
-import net.minecraft.world.World;
 
 @Log4j2
 public class RegisterItems {
@@ -25,59 +17,13 @@ public class RegisterItems {
 
   //TODO : Investigar PlayerItemConsumeEvent para que se ejecute el comando al comer
   //TODO : Investigate Mixing StatusEffects to dont make 7 calls to statusEffects
-  public static final Item INTERTWINED_FATE = new Item(new Item.Settings()
+  public static final Item INTERTWINED_FATE = new CustomItem(new Item.Settings()
       .group(ItemGroup.MISC)
-      .food(new FoodComponent.Builder().hunger(1).saturationModifier(1f).alwaysEdible().build())) {
-    @Override
-    public TypedActionResult<ItemStack> use(World world, PlayerEntity user, Hand hand) {
-      if (this.isFood()) {
-        ItemStack itemStack = user.getStackInHand(hand);
-        if (user.canConsume(this.getFoodComponent().isAlwaysEdible())) {
-          user.setCurrentHand(hand);
-          if (FabricLoader.getInstance().getEnvironmentType() == EnvType.CLIENT) {
-            try {
-              NetworkClient.gachaSend(GachaUtils.EXPENSIVE_GACHA_REQUIEREMENT);
-            } catch (Exception e) {
-              log.fatal("Exception onInitialize at testNetwork() - Message : {}",
-                  e.getMessage());
-            }
-          }
-          return TypedActionResult.consume(itemStack);
-        } else {
-          return TypedActionResult.fail(itemStack);
-        }
-      } else {
-        return TypedActionResult.pass(user.getStackInHand(hand));
-      }
-    }
-  };
+      .food(new FoodComponent.Builder().hunger(1).saturationModifier(1f).alwaysEdible().build()));
 
-  public static final Item ACQUAINT_FATE = new Item(new Item.Settings()
+  public static final Item ACQUAINT_FATE = new CustomItem(new Item.Settings()
       .group(ItemGroup.MISC)
-      .food(new FoodComponent.Builder().hunger(1).saturationModifier(1f).alwaysEdible().build())) {
-    @Override
-    public TypedActionResult<ItemStack> use(World world, PlayerEntity user, Hand hand) {
-      if (this.isFood()) {
-        ItemStack itemStack = user.getStackInHand(hand);
-        if (user.canConsume(this.getFoodComponent().isAlwaysEdible())) {
-          user.setCurrentHand(hand);
-          if (FabricLoader.getInstance().getEnvironmentType() == EnvType.CLIENT) {
-            try {
-              NetworkClient.gachaSend(GachaUtils.CHEAP_GACHA_REQUIEREMENT);
-            } catch (Exception e) {
-              log.fatal("Exception onInitialize at testNetwork() - Message : {}",
-                      e.getMessage());
-            }
-          }
-          return TypedActionResult.consume(itemStack);
-        } else {
-          return TypedActionResult.fail(itemStack);
-        }
-      } else {
-        return TypedActionResult.pass(user.getStackInHand(hand));
-      }
-    }
-  };
+      .food(new FoodComponent.Builder().hunger(1).saturationModifier(1f).alwaysEdible().build()));
 
   public static final Item ADEPTUS_TEMPTATION = new Item(new Item.Settings()
       .group(ItemGroup.MISC)
